rename to `spin kube` to match the SpinKube org
