+++ .github/workflows/test.yml
@@ -26,7 +26,7 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.SPINKUBE_GITHUB_ACCESS_TOKEN }}
 
     - name: Build
-      run: go build -o spin-plugin-k8s main.go
+      run: go build -o spin-plugin-kube main.go
 
     - name: Run tests
       run: go test ./...
