+++ .goreleaser.yml
@@ -1,11 +1,11 @@
 builds:
-- id: spin-plugin-k8s
+- id: spin-plugin-kube
   main: ./
-  binary: k8s
+  binary: kube
   env:
   - CGO_ENABLED=0
   ldflags:
-  - "-s -w -X github.com/spinkube/spin-plugin-k8s/pkg/cmd.Version={{.Version}}"
+  - "-s -w -X github.com/spinkube/spin-plugin-kube/pkg/cmd.Version={{.Version}}"
   goos:
   - darwin
   - linux
@@ -19,7 +19,7 @@ snapshot:
 
 archives:
 - builds:
-  - spin-plugin-k8s
+  - spin-plugin-kube
   name_template: "{{ .ProjectName }}-{{ .Version }}-{{ .Os }}-{{ .Arch }}{{ if .Arm }}v{{ .Arm }}{{ end }}"
   wrap_in_directory: false
   format: tar.gz
