+++ Makefile
@@ -2,7 +2,7 @@ COMMIT = $(shell git rev-parse --short HEAD)
 
 .PHONY: build
 build:
-	go build -ldflags "-X github.com/spinkube/spin-plugin-k8s/pkg/cmd.Version=git-${COMMIT}" -o bin/spin-plugin-k8s .
+	go build -ldflags "-X github.com/spinkube/spin-plugin-kube/pkg/cmd.Version=git-${COMMIT}" -o bin/spin-kube .
 
 .PHONY: install
 install:
