+++ spin-pluginify.toml
@@ -1,7 +1,7 @@
-name = "k8s"
+name = "kube"
 description = "Commands for publishing applications to Kubernetes using the spin-operator."
 version = "0.2.0"
 spin_compatibility = ">=2.0"
 license = "Apache-2.0"
-homepage = "https://github.com/spinkube/spin-plugin-k8s"
-package = "./bin/spin-plugin-k8s"
+homepage = "https://github.com/spinkube/spin-plugin-kube"
+package = "./bin/spin-kube"
